Restructure consensus with FPC enabled by default for 50x speedup
- Implement clean module structure: photon, prism, wave, focus, nova, horizon, flare, nebula, quasar
- Enable FPC by default in all configurations for owned-object transaction speedup
- Wire up engines (chain/dag/pq) with FPC integration
- Use context.Context everywhere for consistency
- Remove all Snow references, replaced with new module names
- BadgerDB v4 as storage backend through luxfi/database

+++ focus/focus.go
@@ -22,10 +22,10 @@ type FocusCounter struct {
 }
 
 // New creates a new focus counter with given beta
-func New(beta int) Confidence {
+func New(beta uint32) Confidence {
 	return &ConfidenceTracker{
-		counter: &FocusCounter{beta: uint32(beta)},
-		beta:    uint32(beta),
+		counter: &FocusCounter{beta: beta},
+		beta:    beta,
 	}
 }
